Fixed entrypoint argus parsing bugs, and added unit tests for CLI entrypoint

+++ src/main_test_unit.py
@@ -1,6 +1,7 @@
 import logging
 import pathlib
 import shutil
+import subprocess
 
 from . import main as test_module
 
@@ -48,4 +49,25 @@ class Test_run:
             f_path=project_dir.joinpath("testing_module").joinpath("test"),
             expected="Testing\n",
         )
+        shutil.rmtree(cache_dir)
+
+    def test_main_entrypoint(self) -> None:
+        cache_dir = TEST_ASSETS_DIR.joinpath(".test_cache")
+        if cache_dir.is_dir():
+            shutil.rmtree(cache_dir)
+
+        p = subprocess.Popen(
+            [
+                "python3",
+                pathlib.Path("src", "main.py"),
+                "--template",
+                TEST_ASSETS_DIR.joinpath("run_case_1"),
+                "--cache",
+                cache_dir,
+            ]
+        )
+        res, _ = p.communicate()
+        p.wait()
+        if p.returncode != 0:
+            raise RuntimeError(res.decode("utf-8"))
         shutil.rmtree(cache_dir)
